chore: optimize non-critical components loading
